move to entry editor tab
