+++ jabgui/src/main/resources/org/jabref/gui/preferences/general/GeneralTab.fxml
@@ -117,7 +117,6 @@
             <Insets top="-8.0"/>
         </padding>
     </HBox>
-    <CheckBox fx:id="enableMscKeywordDescriptions" text="%Enable MSC keyword descriptions"/>
 
     <HBox alignment="CENTER_LEFT" spacing="10.0">
         <CheckBox fx:id="createBackup" text="%Create backup"/>
